Added a PostRiskCalculator [skip hazardlib][demos]
